feat: better obj loading (automatic triangulation), random vert coloring

+++ src/renderer.rs
@@ -241,7 +241,7 @@ impl Renderer {
 			render_pass.set_pipeline(&self.render_pipeline);
 
 			render_pass.set_vertex_buffer(0, mesh.vertex_buffer.slice(..));
-			render_pass.set_index_buffer(mesh.index_buffer.slice(..), wgpu::IndexFormat::Uint16);
+			render_pass.set_index_buffer(mesh.index_buffer.slice(..), wgpu::IndexFormat::Uint32);
 			render_pass.set_bind_group(0, &self.bind_group, &[]);
 
 			render_pass.draw_indexed(0..mesh.index_count, 0, 0..1);
